cambio la inyección de dependencias. La gestión de la incidencia recibida se hace desde un SERVICIO, no desde el controlador. https://github.com/Arquisoft/InciDashboard_e3b/issues/31
