+++ src/main/java/es/uniovi/asw/e3b/incidashboard_e3b/controllers/IncidenciasController.java
@@ -1,8 +1,10 @@
 package es.uniovi.asw.e3b.incidashboard_e3b.controllers;
 
-import java.security.Principal;
-import java.util.List;
-
+import es.uniovi.asw.e3b.incidashboard_e3b.entities.Incidence;
+import es.uniovi.asw.e3b.incidashboard_e3b.entities.Operario;
+import es.uniovi.asw.e3b.incidashboard_e3b.services.EstadoService;
+import es.uniovi.asw.e3b.incidashboard_e3b.services.IncidenciasService;
+import es.uniovi.asw.e3b.incidashboard_e3b.services.OperariosService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -11,11 +13,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import es.uniovi.asw.e3b.incidashboard_e3b.entities.Incidence;
-import es.uniovi.asw.e3b.incidashboard_e3b.entities.Operario;
-import es.uniovi.asw.e3b.incidashboard_e3b.services.EstadoService;
-import es.uniovi.asw.e3b.incidashboard_e3b.services.IncidenciasService;
-import es.uniovi.asw.e3b.incidashboard_e3b.services.OperariosService;
+import java.security.Principal;
+import java.util.List;
 
 @Controller
 public class IncidenciasController {
@@ -92,27 +91,6 @@ public class IncidenciasController {
 		return "redirect:/incidencias/list";
 	}
 
-	/**
-	 * Método que recibe incidencias. Se encarga de seleccionar el operario con
-	 * menos incidencias y asignarle la nueva incidencia recibida
-	 * 
-	 * @param incidencia
-	 *            nueva que se acaba de recibir
-	 */
-	public void recieveIncidence(Incidence incidencia) {
-		// asignar incidencia al operario con menos incidencias asignadas
-		List<Operario> operarios = operariosService.findAll();
-		if (operarios.size() > 0) {
-			int op = 0;
-			for (int i = 0; i < operarios.size(); i++) {
-				if (operarios.get(i).getIncidencias().size() < operarios.get(op).getIncidencias().size())
-					op = i;
-			}
-			incidencia.setOperario(operarios.get(op));
-			operarios.get(op).añadirIncidencia(incidencia);
-		}
-		// meter incidencia en base de datos
-		incidenciasService.addIncidencia(incidencia);
-	}
+
 
 }
