+++ src/main/java/es/uniovi/asw/e3b/incidashboard_e3b/listeners/IncidenceListener.java
@@ -1,19 +1,16 @@
 package es.uniovi.asw.e3b.incidashboard_e3b.listeners;
 
-import javax.annotation.ManagedBean;
-
+import es.uniovi.asw.e3b.incidashboard_e3b.controllers.MainController;
+import es.uniovi.asw.e3b.incidashboard_e3b.entities.Incidence;
+import es.uniovi.asw.e3b.incidashboard_e3b.services.OperariosService;
+import es.uniovi.asw.e3b.incidashboard_e3b.util.ParserJsonToIncidence;
 import org.apache.log4j.Logger;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.kafka.support.KafkaHeaders;
-import org.springframework.messaging.handler.annotation.Header;
 
-import es.uniovi.asw.e3b.incidashboard_e3b.controllers.IncidenciasController;
-import es.uniovi.asw.e3b.incidashboard_e3b.controllers.MainController;
-import es.uniovi.asw.e3b.incidashboard_e3b.entities.Incidence;
-import es.uniovi.asw.e3b.incidashboard_e3b.util.ParserJsonToIncidence;
+import javax.annotation.ManagedBean;
 
 @ManagedBean
 public class IncidenceListener {
@@ -26,6 +23,9 @@ public class IncidenceListener {
 	@Autowired
 	private MainController mainController;
 
+	@Autowired
+	private OperariosService operariosService;
+
 	@KafkaListener(topics = "${kafka.topic:incidences}")
 	public void listen(String data) {
 		logger.info("New incidence received: \"" + data + "\" on topic '" + kafkaTopic + "'");
@@ -34,8 +34,7 @@ public class IncidenceListener {
 		// Recoger incidencia recibida y tratarla // comentado todo... comprobar si es
 		// correcto, sino: eliminar comentario
 		Incidence incidencia = ParserJsonToIncidence.JsonToIncidence(new JSONObject(data));
-		IncidenciasController inciController = new IncidenciasController();
-		inciController.recieveIncidence(incidencia);
+		operariosService.recieveIncidence(incidencia);
 	}
 
 }
